Bugfix: duplicated date numbers in historical dates

+++ recall/__init__.py
@@ -1,11 +1,16 @@
 
+import os
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import LoginManager
 
 app = Flask(__name__)
+# Todo: hide secrets from git
 app.secret_key = 'Melodies of Life 34fj34ofi34jfl3'
-if True:
+
+localhost = bool(os.environ.get('FLASK_DEBUG'))
+
+if not localhost:
     SQLALCHEMY_DATABASE_URI = "mysql+mysqlconnector://{username}:{password}@{hostname}/{databasename}".format(
         username="penlect",
         password="melodiesoflife",
@@ -25,17 +30,18 @@ login_manager.init_app(app)
 import recall.views  # Has to be here
 import recall.models
 
-restart = False
-if restart:
+restart = True
+if localhost and restart:
     db.metadata.drop_all(db.engine, tables=[
-        models.MemoData.__table__,
-        models.RecallData.__table__,
-        models.Correction.__table__,
-        models.User.__table__,
-        models.AlmostCorrectWord.__table__,
+        #models.MemoData.__table__,
+        #models.RecallData.__table__,
+        #models.Correction.__table__,
+        #models.User.__table__,
+        #models.AlmostCorrectWord.__table__,
     ])
 db.create_all()
-if restart:
+
+if False:
     u1 = models.User('sm2017', 'secret', 'hej@email.se', 'Daniel Andersson', 'Sweden')
     u2 = models.User('penlect', '123abc', 'email@email.com', 'Kalle Anka', 'Sweden')
     db.session.add(u1)
